Move geocode and cors middleware to middleware directory
