+++ index.js
@@ -8,7 +8,6 @@ var googleAuth = require('google-auth-library');
 var calendar = google.calendar('v3');
 var jwt = require('jsonwebtoken');
 var passport = require('passport');
-var geocoder = require('geocoder');
 var async = require('async');
 var credentials = require('./config.js');
 var userCalendar = require('./external-apis/calendar.js');
@@ -20,94 +19,25 @@ var moment = require('moment');
 var getNewAccessToken = require('./utils/refresh.js');
 var _ = require('underscore');
 
+// FIX: set secret as random uuid?? Will that break jwts on server restarts?
+// Set secret for jwt signing
+app.set('superSecret', 'anything');
+
+// Set middleware for request body parsing
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
   extended: true
 }));
-app.set('superSecret', 'anything');
+
+// Set middleware for cors headers
+require('./middleware/cors.js')(app);
+
+// Set middlware for gecoding request parameters
+require('./middleware/geocode.js')(app, async, _);
+
 
 var port = process.env.PORT || 3000;
 
-/**
- * Cors headers
- */
-app.use(function(req, res, next) {
-
-  res.header('Access-Control-Allow-Credentials', true);
-  res.header('Access-Control-Allow-Origin', "*");
-  res.header('Access-Control-Allow-Headers', 'X-Requested-With, X-Access-Token, X-HTTP-Method-Override, Content-Type, Accept, Authorization');
-  next();
-});
-
-/**
- * Middleware to convert destination address on req.body.destination to longitude and latitude coordinates
- */
-app.use(function(req, res, next) {
-
-
-  if (req.body.destAddress) {
-
-    async.parallel({
-      geocode: function(cb) {
-        geocoder.geocode(req.body.destAddress, function(err, data) {
-
-          if (err || !data.results[0] || isProbablyWrong) {
-            // console.log('error geocoding:', err)
-            cb(true, null);
-          } else {
-
-            var isProbablyWrong = _.contains(data.results[0].types, 'subpremise');
-
-            if (isProbablyWrong) {
-              cb(true, null);
-            } else {
-              // console.log('Geocode results:', data.results[0])
-              var coordinates = data.results[0].geometry;
-              req.body.destination = {};
-              req.body.destination.longitude = coordinates.location.lng.toString();
-              req.body.destination.latitude = coordinates.location.lat.toString();
-              cb(null, true);
-            }
-          }
-        });
-      },
-      reverseGeocode: function(cb) {
-        var latitude = req.body.origin.latitude;
-        var longitude = req.body.origin.longitude;
-        geocoder.reverseGeocode(Number(latitude), Number(longitude), function(err, data) {
-          if (err) {
-            // console.log('error reverse geocoding', err);
-            cb(err, null);
-          } else {
-            // console.log('reverseGeocode results:', data)
-            req.body.origin.address = data.results[0].formatted_address;
-            cb(null, true);
-          }
-        })
-
-      },
-    },
-      function(err, results) {
-        // console.log('geocode final error:', err);
-        // console.log('geocode final results:', results);
-        if (err) {
-          res.status(400);
-          res.end()
-        } else {
-          next()
-
-        }
-    });
-
-  } else {
-    next();
-  }
-
-});
-
-app.get('/temp', function(req, res) {
-  res.send('<!DOCTYPE html><body><a href="/auth/dummy">Authorize</a></body></html>')
-})
 
 var parseGoogleJwt = function(googleJwt) {
   var parts = googleJwt.split('.');
@@ -180,12 +110,15 @@ app.get('/auth/code', function(req, res) {
 
 
 
-// require('./auth-strategies/google-strategy.js')(passport, app, jwt, null, credentials, UserModel);
 app.use('/api', apiRouter);
 app.use('/cal', calRouter);
 
+
+
+
 var api = require('./routes/api.js')(app, apiRouter, null, User, userCalendar, userMap, uber, calendar, googleAuth, credentials);
-var authenticate = require('./routes/authentication')(app, calRouter, jwt);
+
+var authMiddleware = require('./routes/authentication')(app, calRouter, jwt);
 var main = require('./routes/main.js')(app, calRouter, User, userCalendar, calendar, googleAuth, credentials, getNewAccessToken);
 
 app.listen(port, function() {
